+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased
 
+### Feat
+
+- add visual status
+
 ## v0.0.1 (2022-10-13)
 
 ### Feat